rename database to database/redshift-firehose.js
in order to allow for different database backends
